Adding if statement to check if uncollapsed fastq reads exist before running collate_fastq_reads module

+++ scripts/collate_fastq_reads.R
@@ -22,17 +22,19 @@ if (args$out == ""){
   args$out = args$wkdir
 }
 
-sample_meta = read.csv(args$sample_meta)
-uncollapsed_raw_counts = read.csv(args$uncollapsed_raw_counts)
-
-print("Collating fastq reads")
-raw_counts <- collate_fastq_reads(sample_meta,uncollapsed_raw_counts) 
-
-rc_out_file = paste(
-  args$out,
-  'raw_counts.csv',
-  sep='/'
-)
-print(paste("Writing to file: ", rc_out_file))
-write.csv(raw_counts, rc_out_file, row.names=F, quote=F)
-
+if (file.exists(args$uncollapsed_raw_counts)) {
+  sample_meta = read.csv(args$sample_meta)
+  uncollapsed_raw_counts = read.csv(args$uncollapsed_raw_counts)
+  print("Collating fastq reads")
+  raw_counts <- collate_fastq_reads(sample_meta,uncollapsed_raw_counts) 
+  
+  rc_out_file = paste(
+    args$out,
+    'raw_counts.csv',
+    sep='/'
+  )
+  print(paste("Writing to file: ", rc_out_file))
+  write.csv(raw_counts, rc_out_file, row.names=F, quote=F)
+} else {
+  print("Uncollapsed raw counts file not detected. Proceeding with generating filtered counts file.")
+}
